Address CodeRabbit review on PR #35
Correctness
- landlord_review.dart / review_prompt_providers.dart: clamp() returns num,
  so cast to int before indexing a List and before returning from an
  int-typed method.
- sembast_feedback_repository.dart: import dart:collection for firstOrNull.
- listing_result_card.dart: an interpolated key can never match the catalog;
  use the generated photoCount(count) getter.
- router.dart: whitelist /reviews for landlord and staff sessions so
  LandlordReviewsScreen is reachable instead of redirecting home.
- nps_prompt_gate.dart: await appVersionProvider rather than reading .value
  once (an unresolved PackageInfo permanently suppressed the prompt), and
  decide the second-decline opt-out from pre-prompt state so it no longer
  fires one decline early.

Security and data integrity
- storage.rules: split public listing media into `get` (public) and `list`
  (denied) so retired-but-unswept objects cannot be enumerated. Covered by a
  new emulator test for anonymous get/list on published and stale objects.
- feedback.ts: require an active landlord account via requireActiveLandlord,
  and add a shared per-actor submit throttle across every feedback kind on
  top of the existing NPS cooldown.
- public-seo-handler.ts: drop the 24h stale-while-revalidate window that kept
  unpublished and expired listings servable for ~25h, and mark 404/410/405
  and media-error responses no-store.
- media-publication.ts: sweep superseded delivery objects only after the
  Firestore transaction commits, so an aborted publication no longer deletes
  objects the stored document still references.

Performance
- rating-badges.ts: replace the unbounded collection read with cursor
  pagination over the existing (landlordToken, status) index.
- offline_database.dart / remote_pull_gateway.dart: RemoteMergeResult.failed
  was produced but never observed; log it and count it via a new
  PerfNames.remoteMergeFailure counter.

Localization
- Replace manual English pluralization in public_reviews_section.dart with an
  ICU plural key across all four catalogs, with placeholder metadata mirrored
  so the catalog parity test passes.

+++ lib/features/reviews/application/review_prompt_providers.dart
@@ -105,5 +105,5 @@ final myTenanciesForReviewProvider =
 
 int _monthsBetween(DateTime from, DateTime to) {
   final days = to.difference(from).inDays;
-  return days <= 0 ? 1 : (days / 30.44).round().clamp(1, 600);
+  return days <= 0 ? 1 : (days / 30.44).round().clamp(1, 600).toInt();
 }
